Bajaj Chatbot/requirements.txt
langchain
chromadb
streamlit
pandas
PyPDF2
python-dotenv
sentence-transformers
google-generativeai
